addressing windows compiler warnings. see #98

+++ include/c74_lib_misc.h
@@ -57,8 +57,8 @@ namespace min {
                 if (x > m) {
                     if (x > (m*2.0)) {
                         d = x / m;
-                        di = d;				// cast to long
-                        d = d - di;			// cast to double
+                        di = static_cast<long>(d);	// cast to long
+                        d = d - di;					// cast to double
                         if (di % 2) {
                             if (d < 0)
                                 d = -1.0 - d;
@@ -111,8 +111,8 @@ namespace min {
 				else if (x < 0.0) {
                     if (x < (-m)) {
                         d = x / m;
-                        di = d;				// cast to long
-                        d = d - di;			// cast to double
+                        di = static_cast<long>(d);		// cast to long
+                        d = d - di;						// cast to double
                         x = d * m;
                         if (x < 0.0)
                             x += m;
